Add files via upload
Task-1: The Caesar cipher is one of the earliest and simplest encryption techniques in cryptography. It is named after Julius Caesar, who is historically credited with its use for military communications. It uses a shift key for both encryption and decryption, and the same shift key is required for both processes.